refactor(gateway): split hub.go by extracting SeqGen and pcEntry
hub.go was 818 lines mixing Hub core with two independent components.
SeqGen (sequence generator) moved to seq.go, pcEntry (platform
connection writer) moved to platform_writer.go. hub.go reduced to 572
lines. Pure file split, no logic changes.

Ref #66
